Refactored InputMapping to make it more generic
(*SO - Scriptable Object)
Now to use InputMapping it is needed to create class and make it child of a generic class "InputMapSO" specifying 5 enums for input value types (bool, float, vector2, vector3, quaternion).
To actualy change values in InputMapping SO asset it is needed to create child class of the generic InputMapperSO specifying same enums as in "InputMapSO". After creating scripts with concrete implementation of InputMappingSO and InputMapperSO, create SO assets from these scripts and add InputMapSO to the apropriate field in the InputMapperSO asset.
In the concrete implementation of InputMapperSO it will be needed to create all the logic triggering InputMapperSO base class functions - BoolInputTriggered, FloatInputTrigerred, Vector2InputTrigerred, Vector3InputTrigerred and QuaternionInputTrigerred.